perf(cdogs): render into an RGB565 framebuffer
PicosRenderer.framebuf becomes uint16_t* in host byte order, get_target()
returns uint16_t*, blend_pixel and every drawing primitive work in 16-bit,
and render targets lose their ARGB8888 exception.

SDL_RenderPresent collapses to a single drawImageNN call: the framebuffer
is already in the format the display driver consumes, so the per-frame
full-framebuffer conversion loop and its 153_600-byte staging buffer are
both gone.

Reclaims 460_800 bytes here (bkgTgt 153_600 + framebuf 153_600 +
staging 153_600), 1_075_200 across sub-project 2A.

+++ picos_sdl_impl.h
@@ -70,7 +70,7 @@ typedef struct PicosTexture {
 
 /* ── Internal renderer structure ─────────────────────────────── */
 typedef struct PicosRenderer {
-    uint32_t *framebuf;        /* default render target (ARGB8888) */
+    uint16_t *framebuf;        /* default render target (RGB565, host order) */
     int fb_w, fb_h;            /* framebuffer dimensions */
     uint8_t draw_r, draw_g, draw_b, draw_a;
     int draw_blend_mode;
